KVM: PPC: Reserve KVM_CAP_SPAPR_TCE_VFIO capability number
BugLink: http://bugs.launchpad.net/bugs/1686019

Small change on file include/uapi/linux/kvm.h due to conflics with
defines included by MIPS and S390 patches.

This adds a capability number for in-kernel support for VFIO on
SPAPR platform.

The capability will tell the user space whether in-kernel handlers of
H_PUT_TCE can handle VFIO-targeted requests or not. If not, the user space
must not attempt allocating a TCE table in the host kernel via
the KVM_CREATE_SPAPR_TCE KVM ioctl because in that case TCE requests
will not be passed to the user space which is desired action in
the situation like that.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Reviewed-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Paul Mackerras <paulus@ozlabs.org>
(backported from commit 4898d3f49b5b156c33f0ae0f49ede417ab86195e)
Signed-off-by: Jose Ricardo Ziviani <joserz@linux.vnet.ibm.com>
[saf: Adjust context for 4.11]
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>